+++ kernel/arch/i386/idt.c
@@ -1,3 +1,4 @@
+#include <string.h>
 #include <kernel/system.h>
 
 /* Defines an IDT entry */
@@ -49,7 +50,7 @@ void idt_install()
 {
     /* Sets the special IDT pointer up, just like in 'gdt.c' */
     idtp.limit = (sizeof (struct idt_entry) * 256) - 1;
-    idtp.base = &idt;
+    idtp.base = (unsigned int) &idt;
 
     /* Clear out the entire IDT, initializing it to zeros */
     memset(&idt, 0, sizeof(struct idt_entry) * 256);
